Desktop/kamal ai trading new/requirements.txt
black==24.1.0
fastapi==0.95.2
joblib==1.3.2
mypy==1.9.0
openai==1.9.0
prometheus_client==0.15.0
pydantic==2.6.0
pytest==7.4.0
python-dotenv==1.0.1
pyyaml==6.0
rich==13.5.2
ruff==0.24.0
scikit-learn==1.3.2
textual==0.29.0
uvicorn==0.23.0
starlette==0.27.0
pydantic==1.10.15
